backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
httpx
praw
tavily-python
agents
colorama
matplotlib
